Add support for new stream deck
It looks like there may be a new stream deck (hardware) revision,
which uses the same protocol as the stream deck xl (using jpgs
instead of bmp for images).

This commit introduces a new base class for jpg protocol devices
and StreamDeckXL and StreamDeckRev2 are based on that.

Thanks to Patrick Grote for bringing
this to my attention and for your PR.

Co-Authored-By: patrick-dmxc <patrick.grote@dmxcontrol.de>